tig-benchmarker/master/requirements.txt
blake3
dataclasses
hypercorn
numpy
quart
randomname
uvicorn
fastapi
sqlalchemy
psycopg2-binary
requests